% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_complete}
\alias{heat_map_complete}
\title{Indicate when a heatmap is complete and ready for saving}
\usage{
heat_map_complete()
}
\description{
Indicate when a heatmap is complete and ready for saving
}
\examples{
# Save heatmap
heat_map_save("Heatmap.png",
height = 7, 
width = 5)

# Custom layout
heat_map_layout(layout = c(1,2))

# Construct raw heatmap
heat_map(iris[1:10,],
scale = FALSE,
title = "Iris Raw Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Construct scaled heatmap
heat_map(iris[1:10,],
scale = "range",
title = "Iris Scaled Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Signal completion
heat_map_complete()

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
