% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_custom}
\alias{heat_map_custom}
\title{Create a custom heatmap plot layout}
\usage{
heat_map_custom(popup = TRUE, popup_size = c(8, 8), layout = NULL, ...)
}
\arguments{
\item{popup}{logical indicating whether a popup graphics device should be
opened.}

\item{popup_size}{indicates the size of the popup graphics device in inches,
set to \code{c(8,8)} by default.}

\item{layout}{either a vector of the form c(nrow, ncol) defining the
dimensions of the plot or a matrix defining a more sophisticated layout
(see \code{\link[graphics]{layout}}). Vectors can optionally contain a
third element to indicate whether plots should be placed in row (1) or
column (2) order, set to row order by default.}

\item{...}{additional arguments passed to \code{heat_map_new()}.}
}
\description{
\code{heat_map_custom()} is similar to \code{cheat_map_save()} with the
exception that it doesn't write the plot to a file. \code{heat_map_custom()}
opens a new graphics device and sets the desired layout in preparation for
the addition of heatmaps and other plot objects. Once the custom plot is full
users MUST run \code{cyto_plot_complete()} to close the graphics device and
reset any \code{heat_map()} related settings (see example).
}
\examples{
heat_map_custom(
  popup = FALSE,
  layout = c(1,2)
)
heat_map(
  mtcars
)
plot(
  mtcars[, 1:2],
  pch = 16
)
heat_map_complete()

}
\seealso{
\code{\link{heat_map_save()}}

\code{\link{heat_map_new()}}

\code{\link{heat_map_complete()}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
