% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_highlight}
\alias{heat_map_highlight}
\title{Highlight cells in heatmap}
\usage{
heat_map_highlight()
}
\description{
Highlight cells in heatmap
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
