% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_layout}
\alias{heat_map_layout}
\title{Arrange multiple heatmaps}
\usage{
heat_map_layout()
}
\description{
Arrange multiple heatmaps
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
