% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_record}
\alias{heat_map_record}
\title{Record a custom heatmap}
\usage{
heat_map_record()
}
\description{
Record custom heatmap layout on current graphics device and save to an R
object for future use.
}
\examples{
# Heatmap layout
heat_map_layout(c(1,2))

# Construct raw heatmap
heat_map(iris[1:10,],
scale = FALSE,
title = "Iris Raw Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Construct scaled heatmap
heat_map(iris[1:10,],
scale = "range",
title = "Iris Scaled Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Record heatmap layout
heat_map <- heat_map_record()

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
