% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_reset}
\alias{heat_map_reset}
\title{Reset all heatmap related settings}
\usage{
heat_map_reset()
}
\value{
No return value, called for side effects.
}
\description{
Reset all heatmap related settings
}
\examples{
\donttest{
# Reset HeatmapR settings
heat_map_reset()
}
}
\author{
Dillon Hammill (dillon.hammill21@gmail.com)
}
