% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_save}
\alias{heat_map_save}
\title{Save high resolution images}
\usage{
heat_map_save(
  save_as,
  width = 7,
  height = 7,
  units = "in",
  res = 300,
  multiple = FALSE,
  layout = NULL,
  ...
)
}
\arguments{
\item{save_as}{name of the file to which the plot should be saved (including
the file extension). Supported file formats include png, tiff, jpeg, svg
and pdf.}

\item{width}{numeric indicating the width of exported plot in \code{units},
set to 7 by default for image with width of 7 inches.}

\item{height}{numeric indicating the height of the exported plot in
\code{units}, set to 7 by default for image with height of 7 inches.}

\item{units}{units to be used to set plot size, can be either pixels
(\code{px}), inches (\code{inches}), centimetres (\code{cm}) or millimetres
(\code{mm}). Set to \code{"in"} by default. Units cannot be altered for
\code{svg} and \code{pdf} graphics devices.}

\item{res}{resolution in ppi, set to 300 by default.}

\item{multiple}{logical indicating whether multiple pages should be saved to
separate numbered files, set to \code{TRUE} by default.}

\item{layout}{a vector or matrix defining the custom layout of the plot to be
created using \code{heat_map_layout}, set to NULL by default.}

\item{...}{additional arguments for the appropriate \code{png()},
\code{tiff()}, \code{jpeg()}, \code{svg()} or \code{pdf} graphics devices.}
}
\description{
Save high resolution images
}
\examples{


}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
