% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heap_map-scale.R
\name{heat_map_scale}
\alias{heat_map_scale}
\title{Scale numeric data prior to constructing heatmap}
\usage{
heat_map_scale(x, method = "range")
}
\arguments{
\item{x}{matrix-like object to be scaled.}

\item{method}{type of scaling to perform, can be either \code{'range'},
\code{'mean'} or {'zscore'}. Range scaling normalizes the data to have
limits between 0 and 1. Mean scaling subtracts the mean (calculated
excluding missing values) from each value. Z-score scaling subtracts the
mean from each value and then divides the result by the standard deviation.}

\item{scale}{indicates whether scaling should be applied as per
\code{format}, set to TRUE by default.}
}
\description{
Apply column-wise scaling to numeric columns in a matrix or data.frame prior
to constructing a heatmap.
}
\examples{
# RANGE SCALED DATA
mtcars_scaled <- heat_map_scale(mtcars)
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
