% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map.R
\name{heat_map}
\alias{heat_map}
\title{Create heatmap using base graphics}
\usage{
heat_map(
  x,
  scale = NULL,
  axis_text_x = NULL,
  axis_text_x_side = "bottom",
  axis_text_x_font = 1,
  axis_text_x_size = 1,
  axis_text_x_col = "black",
  axis_text_x_adjust = 1,
  axis_text_x_angle = 0,
  axis_label_x = NULL,
  axis_label_x_font = 2,
  axis_label_x_size = 1.2,
  axis_label_x_col = "black",
  axis_text_y = NULL,
  axis_text_y_side = "left",
  axis_text_y_font = 1,
  axis_text_y_size = 1,
  axis_text_y_col = "black",
  axis_text_y_adjust = 1,
  axis_text_y_angle = 0,
  axis_label_y = NULL,
  axis_label_y_font = 2,
  axis_label_y_size = 1.2,
  axis_label_y_col = "black",
  title = NULL,
  title_text_font = 2,
  title_text_size = 1.5,
  title_text_col = "black",
  box_col_scale = c("red", "black", "green"),
  box_col_alpha = 1,
  box_border_line_type = 1,
  box_border_line_width = 1,
  box_border_line_col = "black",
  margins = NULL,
  legend = TRUE
)
}
\arguments{
\item{x}{matrix or data.frame containing numeric columns to be included in
heatmap. All non-numeric columns will be removed prior to constructing the
heatmap.}
}
\description{
Create heatmap using base graphics
}
\examples{
# Heatmap - Raw Values
heat_map(iris[1:10, ],
  title = "Iris Heatmap",
  axis_label_x = "Plant Parameter",
  axis_label_y = "Row ID"
)

# Heatmap - Scaled
heat_map(iris[1:10, ],
  scale = "range",
  title = "Iris Heatmap",
  axis_label_x = "Plant Parameter",
  axis_label_y = "Row ID",
  box_col_scale = c("yellow", "orange", "red")
)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
