% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map_clust.R
\name{heat_map_clust}
\alias{heat_map_clust}
\title{Perform hierarchical clustering for heatmap}
\usage{
heat_map_clust(
  x,
  tree = "row",
  dist = "euclidean",
  method = "complete",
  scale = FALSE,
  cut = NULL,
  ...
)
}
\arguments{
\item{x}{matrix-like object to cluster. The distance matrix will be computed
using \code{dist} and passed to \code{hclust} for hierarchical clustering.}

\item{tree}{indicates whether hierarchical clustering should be performed by
\code{"row"} or \code{"column"}. Optionally a custom object of class
\code{"dist"} or \code{"hclust"} which will be updated with cluster labels
as specified by \code{cut}.}

\item{dist}{method passed to \code{\link{dist}} to compute distance matrix,
set to \code{"euclidean"} by default.}

\item{method}{agglomeration method passed to \code{\link{hclust}} to perform
hierarchical clustering, set to \code{"complete"} by default.}

\item{scale}{logical indicating whether branch heights should be scaled for
better visualisation, set to FALSE by default.}

\item{cut}{value less one specifying the tree cutpoint as a proportion of the
tree height or a value greater than 1 indicating the number of desired
clusters.}

\item{...}{additional arguments passed to \code{\link{dist}} or
\code{\link{hclust}}.}
}
\value{
object of class \code{hclust} which describes the tree produced by
the clustering process.
}
\description{
Perform hierarchical clustering for heatmap
}
\examples{
# Hierarchical clustering
heat_map_hclust <- heat_map_clust(mtcars, cut = 5)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
