% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map_clust.R
\name{heat_map_clust}
\alias{heat_map_clust}
\title{Perform hierarchical clustering for heatmap}
\usage{
heat_map_clust(x, dist_method = "euclidean", clust_method = "complete", ...)
}
\arguments{
\item{x}{matrix-like object to cluster. The distance matrix will be computed
using \code{dist} and passed to \code{hclust} for hierarchical clustering.}

\item{...}{additional arguments passed to \code{\link{dist}} or
\code{\link{hclust}}.}
}
\value{
object of class \code{hclust} which describes the tree produced by
the clustering process.
}
\description{
Perform hierarchical clustering for heatmap
}
\examples{

# Hierarchical clustering
heat_map_hclust <- heat_map_clust(mtcars)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
