% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_complete}
\alias{heat_map_complete}
\title{Indicate when a heatmap is complete and ready for saving}
\usage{
heat_map_complete()
}
\description{
Indicate when a heatmap is complete and ready for saving
}
\examples{
# Save Heatmap
heat_map_save("Heatmap.png",
height = 7, 
width = 5)

# Construct Heatmap
heat_map(iris[1:10,],
scale = "range",
title = "Iris Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Signal Completion
heat_map_complete()

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
