% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_layout}
\alias{heat_map_layout}
\title{Arrange multiple heatmaps}
\usage{
heat_map_layout(layout = NULL)
}
\arguments{
\item{layout}{either a vector of the form c(nrow, ncol) defining the
dimensions of the plot or a matrix defining a more sophisticated layout
(see \code{\link[graphics]{layout}}). Vectors can optionally contain a
third element to indicate whether plots should be placed in row (1) or
column (2) order, set to row order by default.}
}
\description{
Arrange multiple heatmaps
}
\examples{
\dontrun{
# Save heatmap
heat_map_save("Heatmap.png",
height = 7, 
width = 15)

# Custom layout
heat_map_layout(layout = c(1,2))

# Construct raw heatmap
heat_map(iris[1:10,],
scale = FALSE,
title = "Iris Raw Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Construct scaled heatmap
heat_map(iris[1:10,],
scale = "range",
title = "Iris Scaled Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Signal completion
heat_map_complete()
}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
