% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_new}
\alias{heat_map_new}
\title{Open a pop-up graphics device for heatmaps}
\usage{
heat_map_new(popup = FALSE, popup_size = c(8, 8), ...)
}
\arguments{
\item{popup}{logical indicating whether a popup graphics device should be
opened.}

\item{popup_size}{indicates the size of the popup graphics device in inches,
set to \code{c(8,8)} by default.}

\item{...}{additional arguments passed to
\code{\link[grDevices:dev]{dev.new}}:}
}
\description{
Open a pop-up graphics device for heatmaps
}
\examples{
\dontrun{
# Open platform-specific graphics device
heat_map_new(popup = TRUE)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
