% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_record}
\alias{heat_map_record}
\title{Record a custom heatmap}
\usage{
heat_map_record()
}
\description{
Record a custom heatmap
}
\examples{
# Heatmap layout
heat_map_layout(1,2)

# Construct Raw Heatmap
heat_map(iris[1:10,],
scale = NULL,
title = "Iris Raw Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Construct Scaled Heatmap
heat_map(iris[1:10,],
scale = "range",
title = "Iris Scaled Heatmap",
axis_label_x = "Plant Parameter",
axis_label_y = "Row ID")

# Record Heatmap layout
heat_map <- heat_map_record()

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
