% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map-helpers.R
\name{heat_map_reset}
\alias{heat_map_reset}
\title{Reset all heatmap related settings}
\usage{
heat_map_reset()
}
\description{
Reset all heatmap related settings
}
\examples{
# Reset heatmapr settings
heat_map_reset()

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
