% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heap_map-scale.R
\name{heat_map_scale}
\alias{heat_map_scale}
\title{Scale numeric data prior to constructing heat_map}
\usage{
heat_map_scale(x, scale = "column", method = "range")
}
\arguments{
\item{x}{matrix-like object to be scaled.}

\item{scale}{indicates whether the data should be scaled by \code{"row"} or
\code{"column"}, set to \code{"column"} by default.}

\item{method}{type of scaling to perform, can be either \code{'range'},
\code{'mean'} or {'zscore'}. Range scaling normalizes the data to have
limits between 0 and 1. Mean scaling subtracts the mean (calculated
excluding missing values) from each value. Z-score scaling subtracts the
mean from each value and then divides the result by the standard deviation.}
}
\description{
Apply column-wise or row-wise scaling to numeric columns in a matrix or
data.frame prior to constructing a \code{heat_map}.
}
\examples{
# Range scaling
mtcars_scale_range <- heat_map_scale(mtcars,
method = "range")

# Mean scaling
mtcars_scale_mean <- heat_map_scale(mtcars,
method = "mean")

# Z-score scaling
mtcars_scale_zscore <- heat_map_scale(mtcars,
method = "zscore")

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
